# packages/api-contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Futrob Private API
  version: 0.1.0
  description: Private HTTP API for Futrob clients (web, React Native / Expo,
    scripts). Not a public third-party API.
servers:
  - url: /api/v1
    description: Same Worker as apps/web
security:
  - bearerAuth: []
...
paths:
...
  /identity/onboarding/invitation/preview:
    post:
      operationId: inspectCompetitionInvitation
      tags:
        - onboarding
      summary: Inspect a competition invitation without consuming it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Sanitized competition invitation preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectCompetitionInvitationResponse"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
        "429":
          $ref: "#/components/responses/RateLimited"
      parameters:
        - *a2
...
  /competitions/invitations/accept:
    post:
      operationId: acceptCompetitionInvitation
      tags:
        - competitions
      summary: Accept a competition invitation and join its contextual membership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Competition invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptCompetitionInvitationResponse"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "429":
          $ref: "#/components/responses/RateLimited"
      parameters:
        - *a2
...
  /organizations/{organizationId}/competitions:
    get:
      operationId: listOrganizationCompetitions
      tags:
        - competitions
      summary: List competitions in an organization
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Organization competitions newest-first
          content:
            application/json:
              schema:
                type: object
                required:
                  - competitions
                properties:
                  competitions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Competition"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
    post:
      operationId: createCompetitionDraft
      tags:
        - competitions
      summary: Create a competition draft in an organization
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CompetitionDraftInput"
                - type: object
                  properties:
                    creationKey:
                      type: string
                      pattern: ^[A-Za-z0-9_-]{8,120}$
                      description: "Client key: a retry returns the same draft and reuses the cover
                        upload key"
      responses:
        "201":
          description: Competition draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDraft"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
...
  /organizations/{organizationId}/competitions/{competitionId}:
    get:
      operationId: getCompetitionDraft
      tags:
        - competitions
      summary: Get an organization-scoped competition draft
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Competition and current rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDraft"
          headers:
            X-Request-ID: *a1
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
    patch:
      operationId: updateCompetitionDraft
      tags:
        - competitions
      summary: Update the editable structure and rules of a competition draft
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompetitionDraftInput"
      responses:
        "200":
          description: Saved draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDraft"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
...
  /organizations/{organizationId}/competitions/{competitionId}/participants:
    get:
      operationId: listCompetitionParticipants
      tags:
        - competitions
      summary: List draft participants
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Participants
          content:
            application/json:
              schema:
                type: object
                required:
                  - participants
                properties:
                  participants:
                    type: array
                    items:
                      $ref: "#/components/schemas/CompetitionEntry"
          headers:
            X-Request-ID: *a1
        "403":
          $ref: "#/components/responses/ApiError"
    post:
      operationId: addCompetitionParticipant
      tags:
        - competitions
      summary: Add an approved existing or inline-created Team
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionParticipantInput"
      responses:
        "201":
          description: Approved participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionEntry"
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
  /organizations/{organizationId}/competitions/{competitionId}/participants/{entryId}:
    delete:
      operationId: removeCompetitionParticipant
      tags:
        - competitions
      summary: Remove a participant while the competition is a draft
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
        - name: entryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Participant removed
          headers:
            X-Request-ID: *a1
        "403":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
  /organizations/{organizationId}/competitions/{competitionId}/publish:
    post:
      operationId: publishCompetition
      tags:
        - competitions
      summary: Publish and structurally lock a valid draft with at least two approved
        participants
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published competition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDraft"
          headers:
            X-Request-ID: *a1
        "403":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
...
  /organizations/{organizationId}/competitions/{competitionId}/invitations:
    post:
      operationId: createCompetitionInvitation
      tags:
        - competitions
      summary: Create an invitation scoped to a competition
      parameters:
        - *a2
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
        - name: competitionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum:
                    - staff
                    - captain
                    - player
                email:
                  type: string
                  format: email
                expiresInMs:
                  type: integer
                  minimum: 1
                redeemPolicy:
                  type: string
                  enum:
                    - single
                    - multi
                maxRedemptions:
                  type: integer
                  minimum: 1
                  description: Required when redeemPolicy is multi
      responses:
        "201":
          description: Competition invitation created
          content:
            application/json:
              schema:
                type: object
                required:
                  - invitationId
                  - competitionId
                  - token
                  - expiresAt
                  - redeemPolicy
                  - maxRedemptions
                properties:
                  invitationId:
                    type: string
                  competitionId:
                    type: string
                  token:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                  redeemPolicy:
                    type: string
                    enum:
                      - single
                      - multi
                  maxRedemptions:
                    type: integer
                    minimum: 1
                    nullable: true
          headers:
            X-Request-ID: *a1
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "403":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional in MVP scaffold; required when identity is wired.
...
  headers:
    RequestId:
      description: Correlation UUID for support and operational diagnostics.
      schema:
        type: string
        format: uuid
    RetryAfter:
      description: Seconds until this rate-limited operation can be retried.
      schema:
        type: integer
        minimum: 1
  schemas:
...
    ApiError:
      type: object
      required:
        - code
        - messageKey
      properties:
        code:
          type: string
        messageKey:
          type: string
        requestId:
          type: string
          format: uuid
        retryAfterSeconds:
          type: integer
          minimum: 1
        details:
          type: object
          additionalProperties: false
          properties:
            organizationId:
              type: string
            role:
              type: string
            status:
              type: string
            path:
              type: string
            body: {}
            issues: {}
            externalClubId:
              type: string
            cause: {}
            completedPath:
              type: string
            requestedPath:
              type: string
...
    CompetitionDraftInput:
      type: object
      required:
        - name
        - gameEdition
        - platform
        - region
        - timeZone
        - format
      properties:
        teams:
          $ref: "#/components/schemas/CompetitionTeamRange"
        schedule:
          $ref: "#/components/schemas/CompetitionSchedule"
        cover:
          $ref: "#/components/schemas/CompetitionCover"
        name:
          type: string
          minLength: 1
          maxLength: 120
        gameEdition:
          type: string
          minLength: 1
          maxLength: 40
        platform:
          type: string
          enum:
            - playstation
            - xbox
            - pc
            - nintendo-switch-1
            - nintendo-switch-2
        region:
          type: string
          enum:
            - america
            - south-america
            - north-central-america
            - europe
            - africa
            - asia
            - middle-east
            - oceania
        timeZone:
          type: string
          minLength: 1
          maxLength: 100
        format:
          type: string
          enum:
            - league
            - knockout
            - groups-knockout
            - league-playoffs
...
    Competition:
      type: object
      required:
        - id
        - organizationId
        - name
        - status
        - modality
        - gameEdition
        - platform
        - region
        - timeZone
        - format
        - teams
        - schedule
        - cover
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        organizationId:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - draft
            - registration
            - published
            - paused
            - finished
            - archived
        modality:
          type: string
          const: fc-clubs
        gameEdition:
          type: string
        platform:
          type: string
          enum:
            - playstation
            - xbox
            - pc
            - nintendo-switch-1
            - nintendo-switch-2
        region:
          type: string
          enum:
            - america
            - south-america
            - north-central-america
            - europe
            - africa
            - asia
            - middle-east
            - oceania
        timeZone:
          type: string
        format:
          type: string
          enum:
            - league
            - knockout
            - groups-knockout
            - league-playoffs
        teams:
          $ref: "#/components/schemas/CompetitionTeamRange"
        schedule:
          $ref: "#/components/schemas/CompetitionSchedule"
        cover:
          $ref: "#/components/schemas/CompetitionCover"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CompetitionDraft:
      type: object
      required:
        - competition
        - rules
      properties:
        competition:
          $ref: "#/components/schemas/Competition"
        rules:
          $ref: "#/components/schemas/CompetitionRules"
    UpdateCompetitionDraftInput:
      allOf:
        - $ref: "#/components/schemas/CompetitionDraftInput"
        - type: object
          required:
            - rules
          properties:
            rules:
              type: object
              required:
                - regularStage
                - knockoutStage
                - maxRosterSize
              properties:
                regularStage:
                  anyOf:
                    - $ref: "#/components/schemas/CompetitionMatchRules"
                    - type: "null"
                knockoutStage:
                  anyOf:
                    - $ref: "#/components/schemas/CompetitionMatchRules"
                    - type: "null"
                maxRosterSize:
                  anyOf:
                    - type: integer
                      minimum: 1
                    - type: "null"
...
    CompetitionEntry:
      type: object
      required:
        - id
        - organizationId
        - competitionId
        - teamId
        - status
        - createdAt
      properties:
        id:
          type: string
        organizationId:
          type: string
        competitionId:
          type: string
        teamId:
          type: string
        status:
          type: string
          enum:
            - pending
            - approved
            - rejected
        createdAt:
          type: string
          format: date-time
    CompetitionTeamRange:
      type: object
      required:
        - min
        - max
      description: Approved-team bounds; max null means no limit. Approvals stop at max.
      properties:
        min:
          type: integer
          minimum: 2
          maximum: 256
        max:
          type:
            - integer
            - "null"
          minimum: 2
          maximum: 256
...
    CompetitionParticipantInput:
      oneOf:
        - type: object
          required:
            - kind
            - teamId
          properties:
            kind:
              type: string
              const: existing-team
            teamId:
              type: string
        - type: object
          required:
            - kind
            - name
            - creationKey
          properties:
            kind:
              type: string
              const: new-team
            name:
              type: string
              minLength: 1
              maxLength: 120
            creationKey:
              type: string
              minLength: 1
              maxLength: 160
...
    InspectCompetitionInvitationResponse:
      type: object
      required:
        - organizationId
        - organizationName
        - competitionId
        - competitionName
        - competitionRole
        - expiresAt
      properties:
        organizationId:
          type: string
        organizationName:
          type: string
        competitionId:
          type: string
        competitionName:
          type: string
        competitionRole:
          type: string
          enum:
            - staff
            - captain
            - player
        expiresAt:
          type: string
          format: date-time
    AcceptCompetitionInvitationResponse:
      type: object
      required:
        - organizationId
        - organizationName
        - role
        - competitionId
        - competitionName
        - destination
      properties:
        organizationId:
          type: string
        organizationName:
          type: string
        role:
          type: string
          enum:
            - organizer
            - staff
            - member
        competitionId:
          type: string
        competitionName:
          type: string
        competitionRole:
          type: string
          enum:
            - staff
            - captain
            - player
        destination:
          type: object
          required:
            - kind
            - organizationId
            - competitionId
          properties:
            kind:
              type: string
              const: competition
            organizationId:
              type: string
            competitionId:
              type: string
...
  responses:
    ApiError:
      description: API error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"